📑 :: 습관 엔티티 수정
